rag_queue/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
croniter==6.2.2
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
filelock==3.25.2
filetype==1.2.0
Flask==3.1.3
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.64.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.1
langsmith==0.7.25
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
openai==2.30.0
orjson==3.11.8
packaging==26.0
pandas==3.0.2
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.17.1
redis==7.4.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rq==2.7.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.52.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.8
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
